Initial Economy updates
:cl:
bugfix: Due to an accounting error (the nonexistence of accounting), you were not actually being paid wages. You will now be paid 'wages' for your time on station taken out of Cargo's budget. There's not much to do with them yet, besides buy Laptops, but Laptops are Laptops. Be happy with that, you ungrateful sods.
/:cl: